# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Manager Command Center
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type IssuePriority = (typeof IssuePriority)[keyof typeof IssuePriority];

export const IssuePriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type IssueStatus = (typeof IssueStatus)[keyof typeof IssueStatus];

export const IssueStatus = {
  open: "open",
  in_progress: "in_progress",
  blocked: "blocked",
  resolved: "resolved",
} as const;

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  on_hold: "on_hold",
  completed: "completed",
} as const;

export type ActivityCategory =
  (typeof ActivityCategory)[keyof typeof ActivityCategory];

export const ActivityCategory = {
  client_call: "client_call",
  code_review: "code_review",
  planning: "planning",
  one_on_one: "one_on_one",
  support: "support",
  deployment: "deployment",
  documentation: "documentation",
  other: "other",
} as const;

export interface Project {
  id: number;
  name: string;
  client: string;
  /** @nullable */
  description: string | null;
  status: ProjectStatus;
  /** Hex color or token identifier for visual differentiation */
  color: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  client: string;
  /** @nullable */
  description?: string | null;
  status?: ProjectStatus;
  color?: string;
}

export interface UpdateProjectInput {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  client?: string;
  /** @nullable */
  description?: string | null;
  status?: ProjectStatus;
  color?: string;
}

export interface Member {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  email: string | null;
  avatarColor: string;
  createdAt: string;
}

export interface CreateMemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @nullable */
  email?: string | null;
  avatarColor?: string;
}

export interface UpdateMemberInput {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  role?: string;
  /** @nullable */
  email?: string | null;
  avatarColor?: string;
}

export interface Issue {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  description: string | null;
  priority: IssuePriority;
  status: IssueStatus;
  /** @nullable */
  assigneeId: number | null;
  /**
   * Free-text client or stakeholder name
   * @nullable
   */
  reportedBy: string | null;
  /** @nullable */
  dueDate: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateIssueInput {
  projectId: number;
  /** @minLength 1 */
  title: string;
  /** @nullable */
  description?: string | null;
  priority: IssuePriority;
  status?: IssueStatus;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  reportedBy?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface UpdateIssueInput {
  projectId?: number;
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  priority?: IssuePriority;
  status?: IssueStatus;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  reportedBy?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface Activity {
  id: number;
  /** @nullable */
  projectId: number | null;
  category: ActivityCategory;
  title: string;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  durationMinutes: number | null;
  activityDate: string;
  createdAt: string;
}

export interface CreateActivityInput {
  /** @nullable */
  projectId?: number | null;
  category: ActivityCategory;
  /** @minLength 1 */
  title: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  activityDate: string;
}

export interface UpdateActivityInput {
  /** @nullable */
  projectId?: number | null;
  category?: ActivityCategory;
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  activityDate?: string;
}

export interface DashboardSummary {
  activeProjects: number;
  totalProjects: number;
  teamSize: number;
  openIssues: number;
  criticalIssues: number;
  resolvedThisWeek: number;
  activitiesToday: number;
  minutesLoggedThisWeek: number;
}

export interface PriorityBucket {
  priority: IssuePriority;
  count: number;
}

export interface ProjectHealth {
  projectId: number;
  name: string;
  client: string;
  color: string;
  openIssues: number;
  criticalIssues: number;
  resolvedIssues: number;
}

export type GetRecentActivitiesParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type ListIssuesParams = {
  projectId?: number;
  status?: IssueStatus;
  priority?: IssuePriority;
};

export type ListActivitiesParams = {
  projectId?: number;
  /**
   * Filter activities to a specific YYYY-MM-DD date
   */
  date?: string;
};