# lib/api-zod/src/generated/types/activityCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the Manager Command Center
 * OpenAPI spec version: 0.1.0
 */

export type ActivityCategory =
  (typeof ActivityCategory)[keyof typeof ActivityCategory];

export const ActivityCategory = {
  client_call: "client_call",
  code_review: "code_review",
  planning: "planning",
  one_on_one: "one_on_one",
  support: "support",
  deployment: "deployment",
  documentation: "documentation",
  other: "other",
} as const;